requirements.txt
# ==============================
# Core API & Backend
# ==============================
fastapi>=0.128.0
uvicorn>=0.40.0
pydantic>=2.12.0
pydantic-settings>=2.12.0
python-dotenv>=1.2.0
python-multipart>=0.0.22

# ==============================
# Database & Vector Storage
# ==============================
sqlalchemy>=2.0.46
psycopg2-binary>=2.9.11
pgvector>=0.4.2
chromadb>=0.4.24

# ==============================
# Document Processing
# ==============================
pypdf>=6.6.2
pdfplumber>=0.10.3
python-docx>=0.8.11
pdf2image>=1.16.3
pillow>=12.1.0
pytesseract>=0.3.10

# ==============================
# NLP & LLMs (Python 3.11 safe)
# ==============================
spacy>=3.7.2
transformers>=4.40.0,<4.44.0
sentence-transformers>=2.7.0
tokenizers>=0.15.0
torch>=2.1.0

# ==============================
# Arabic NLP (FIXED)
# ==============================
camel-tools>=1.5.7
regex>=2026.1.15

# ==============================
# RAG & LLM Orchestration
# ==============================
langchain>=0.3.7
langchain-core>=0.3.15
langchain-community>=0.3.1
langchain-text-splitters>=0.3.2
langsmith>=0.6.7

# ==============================
# Embeddings APIs (optional)
# ==============================
openai>=2.16.0
anthropic>=0.77.0

# ==============================
# Utilities
# ==============================
numpy>=1.26.0
pandas>=2.1.0
orjson>=3.9.0
tqdm>=4.66.0
loguru>=0.7.2
tenacity>=8.2.0
requests>=2.31.0

# ==============================
# UI / Debug
# ==============================
streamlit>=1.32.0
rich>=13.7.0
